+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Examples/PoolLifetime.hs
@@ -34,6 +34,7 @@ import Cardano.Ledger.Coin (
   CompactForm (CompactCoin),
   DeltaCoin (..),
   addDeltaCoin,
+  knownNonZeroCoin,
   toDeltaCoin,
  )
 import Cardano.Ledger.Compactible
@@ -109,6 +110,7 @@ import Test.Cardano.Ledger.Shelley.Generator.Core (
  )
 import Test.Cardano.Ledger.Shelley.Generator.EraGen (genesisId)
 import Test.Cardano.Ledger.Shelley.Generator.ShelleyEraGen ()
+import Test.Cardano.Ledger.Shelley.Rewards (mkSnapShot)
 import Test.Cardano.Ledger.Shelley.Rules.Chain (ChainState (..))
 import Test.Cardano.Ledger.Shelley.Utils (
   epochSize,
@@ -385,18 +387,19 @@ blockEx3 =
 
 snapEx3 :: SnapShot
 snapEx3 =
-  SnapShot
-    { ssStake =
-        mkStake
-          [ (Cast.aliceSHK, aliceCoinEx2Base <> aliceCoinEx2Ptr)
-          , (Cast.bobSHK, bobInitCoin)
-          ]
-    , ssDelegations =
-        [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
-        , (Cast.bobSHK, aikColdKeyHash Cast.alicePoolKeys)
+  let
+    stake =
+      mkStake
+        [ (Cast.aliceSHK, aliceCoinEx2Base <> aliceCoinEx2Ptr)
+        , (Cast.bobSHK, bobInitCoin)
         ]
-    , ssPoolParams = [(aikColdKeyHash Cast.alicePoolKeys, Cast.aliceStakePoolParams)]
-    }
+    delegations =
+      [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
+      , (Cast.bobSHK, aikColdKeyHash Cast.alicePoolKeys)
+      ]
+    poolParams = [(aikColdKeyHash Cast.alicePoolKeys, Cast.aliceStakePoolParams)]
+   in
+    mkSnapShot stake delegations poolParams
 
 expectedStEx3 :: ChainState ShelleyEra
 expectedStEx3 =
@@ -522,20 +525,21 @@ blockEx5 =
 
 snapEx5 :: SnapShot
 snapEx5 =
-  SnapShot
-    { ssStake =
-        mkStake
-          [ (Cast.aliceSHK, aliceCoinEx4Base <> aliceCoinEx2Ptr)
-          , (Cast.carlSHK, carlMIR)
-          , (Cast.bobSHK, bobInitCoin)
-          ]
-    , ssDelegations =
-        [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
-        , (Cast.carlSHK, aikColdKeyHash Cast.alicePoolKeys)
-        , (Cast.bobSHK, aikColdKeyHash Cast.alicePoolKeys)
+  let
+    stake =
+      mkStake
+        [ (Cast.aliceSHK, aliceCoinEx4Base <> aliceCoinEx2Ptr)
+        , (Cast.carlSHK, carlMIR)
+        , (Cast.bobSHK, bobInitCoin)
         ]
-    , ssPoolParams = [(aikColdKeyHash Cast.alicePoolKeys, Cast.aliceStakePoolParams)]
-    }
+    delegations =
+      [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
+      , (Cast.carlSHK, aikColdKeyHash Cast.alicePoolKeys)
+      , (Cast.bobSHK, aikColdKeyHash Cast.alicePoolKeys)
+      ]
+    poolParams = [(aikColdKeyHash Cast.alicePoolKeys, Cast.aliceStakePoolParams)]
+   in
+    mkSnapShot stake delegations poolParams
 
 pdEx5 :: PoolDistr
 pdEx5 =
@@ -548,7 +552,7 @@ pdEx5 =
             Cast.aliceVRFKeyHash
         )
     )
-    (CompactCoin 1)
+    (knownNonZeroCoin @1)
 
 expectedStEx5 :: ChainState ShelleyEra
 expectedStEx5 =
@@ -778,14 +782,17 @@ blockEx9 =
 
 snapEx9 :: SnapShot
 snapEx9 =
-  snapEx5
-    { ssStake =
-        mkStake
-          [ (Cast.bobSHK, bobInitCoin <> bobRAcnt8)
-          , (Cast.aliceSHK, aliceCoinEx4Base <> aliceCoinEx2Ptr <> aliceRAcnt8)
-          , (Cast.carlSHK, carlMIR)
-          ]
-    }
+  let
+    stake =
+      mkStake
+        [ (Cast.bobSHK, bobInitCoin <> bobRAcnt8)
+        , (Cast.aliceSHK, aliceCoinEx4Base <> aliceCoinEx2Ptr <> aliceRAcnt8)
+        , (Cast.carlSHK, carlMIR)
+        ]
+    delegations = ssDelegations snapEx5
+    poolParams = ssPoolParams snapEx5
+   in
+    mkSnapShot stake delegations poolParams
 
 expectedStEx9 :: ChainState ShelleyEra
 expectedStEx9 =
@@ -999,17 +1006,19 @@ blockEx12 =
 
 snapEx12 :: SnapShot
 snapEx12 =
-  snapEx9
-    { ssStake =
-        mkStake
-          [ (Cast.aliceSHK, aliceRAcnt8 <> aliceCoinEx2Ptr <> aliceCoinEx11Ptr)
-          , (Cast.carlSHK, carlMIR)
-          ]
-    , ssDelegations =
-        [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
-        , (Cast.carlSHK, aikColdKeyHash Cast.alicePoolKeys)
+  let
+    stake =
+      mkStake
+        [ (Cast.aliceSHK, aliceRAcnt8 <> aliceCoinEx2Ptr <> aliceCoinEx11Ptr)
+        , (Cast.carlSHK, carlMIR)
         ]
-    }
+    delegations =
+      [ (Cast.aliceSHK, aikColdKeyHash Cast.alicePoolKeys)
+      , (Cast.carlSHK, aikColdKeyHash Cast.alicePoolKeys)
+      ]
+    poolParams = ssPoolParams snapEx9
+   in
+    mkSnapShot stake delegations poolParams
 
 expectedStEx12 :: ChainState ShelleyEra
 expectedStEx12 =