+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -234,6 +234,8 @@ instance ToExpr SnapShots
 
 instance ToExpr SnapShot
 
+instance ToExpr StakePoolSnapShot
+
 deriving newtype instance ToExpr Stake
 
 instance ToExpr (PState era)