+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Generic/Trace.hs
@@ -20,8 +20,8 @@ module Test.Cardano.Ledger.Generic.Trace where
 import Cardano.Ledger.Address (Addr (..))
 import Cardano.Ledger.Alonzo.Rules (AlonzoUtxowPredFailure (..))
 import Cardano.Ledger.Babbage.Rules (BabbageUtxowPredFailure (..))
-import Cardano.Ledger.BaseTypes (BlocksMade (..), Globals (networkId))
-import Cardano.Ledger.Coin (CompactForm (CompactCoin))
+import Cardano.Ledger.BaseTypes (BlocksMade (..), Globals)
+import Cardano.Ledger.Coin (knownNonZeroCoin)
 import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (
   EpochState (..),
@@ -105,7 +105,9 @@ import Test.Control.State.Transition.Trace.Generator.QuickCheck (HasTrace (..), 
 genRsTxAndModel ::
   forall era.
   EraGenericGen era =>
-  Int -> SlotNo -> GenRS era (Tx TopTx era)
+  Int ->
+  SlotNo ->
+  GenRS era (Tx TopTx era)
 genRsTxAndModel n slot = do
   (_, tx) <- genAlonzoTx slot
   modifyModel (\model -> applyTx n slot model tx)
@@ -166,7 +168,7 @@ initialMockChainState gstate =
         , nesBcur = BlocksMade Map.empty
         , nesEs = makeEpochState gstate ledgerstate
         , nesRu = SNothing
-        , nesPd = PoolDistr (gsInitialPoolDistr gstate) (CompactCoin 1)
+        , nesPd = PoolDistr (gsInitialPoolDistr gstate) (knownNonZeroCoin @1)
         , stashedAVVMAddresses = stashedAVVMAddressesZero (reify @era)
         }
 
@@ -192,7 +194,7 @@ snaps (LedgerState UTxOState {utxosUtxo = u, utxosFees = f} certState) =
   where
     pstate = certState ^. certPStateL
     dstate = certState ^. certDStateL
-    snap = stakeDistr (networkId testGlobals) u dstate pstate
+    snap = stakeDistr u dstate pstate
 
 -- ==============================================================================
 
@@ -456,7 +458,7 @@ forAllTraceFromInitState baseEnv maxTraceLength traceGenEnv genSt0 =
 --   Under the assumption that shorter tests have advantages
 --   like not getting turned off because the tests take too long. A glaring failure is
 --   likely to be caught in 'n' tests, rather than the standard '100'
-testPropMax :: Testable prop => Int -> String -> prop -> Spec
+testPropMax :: (HasCallStack, Testable prop) => Int -> String -> prop -> Spec
 testPropMax n name x = prop name (withMaxSuccess n x)
 
 chainTest ::