+++ eras/shelley/impl/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.18.0.0
 
+* Add `calcNonMyopicMemberReward` and deprecate `nonMyopicMemberRew` in its favor.
+* Add `calcStakePoolDesirability` and deprecate `desirability` in its favor.
+* Change type signature of `mkPoolRewardInfo`, `getTopRankedPools`
 * Re-export `AccountAddress`, `AccountId`, and pattern synonym `RewardAccount` from `Cardano.Ledger.Address`
 * Move delegatee registration validation from `DELEGS` to `DELEG` rule:
   - Add `DelegateeNotRegisteredDELEG` to `ShelleyDelegPredFailure`